feat: tela de deleção de um usuário implementada

+++ client/src/ui/screens/SpecificUser/index.tsx
@@ -52,7 +52,7 @@ const SpecificUser = () => {
   const deleteUser = async () => {
     if (!id) return;
     try {
-      await api.delete(`/users/${id}`);
+      await api.delete(`/users/admin/remove/${id}`);
       navigate("/users");
     } catch (error) {
       console.error("Erro ao excluir usuário:", error);
